# Optimize read locker cleanup