@harshavardhana 2022-01-27
> Looks like some legit issue at dsync
> 
> ```
> --- FAIL: TestUnlockShouldNotTimeout (60.00s)
>     dsync_test.go:286: Unlock timed out, which should not happen
> FAIL
> FAIL	github.com/minio/minio/internal/dsync	156.012s
> ```

Looks like a transient issue but it may be something to think about.